refactor: settings파일을 config 디렉터리로 재구성
- 더 나은 명명 규칙을 위해 `setting.ts`의 이름을 `settings.ts`로 변경합니다.
- settings.ts` 파일을 새 `config` 디렉터리로 이동합니다.
- main.ts`의 가져오기 문을 업데이트하여 새 `config/settings.ts` 경로에서 가져오도록 합니다.
